Ansys analysis updates
