+++ examples/myBlade_loadsTable.json
@@ -1,0 +1,1 @@
+{"rBlade": [2.5, 7.5, 12.5, 17.5, 22.5, 27.5, 32.5, 37.5, 42.5, 47.5, 52.5, 57.5, 62.5, 67.5, 72.5, 77.5, 82.5, 87.5, 92.5, 97.5], "Fxb": [136110.0, -129290.0, 138140.0, -97060.0, 134640.0, -57130.0, 126980.0, -61170.0, 116710.0, -33800.0, 102780.0, -1900.0, 83630.0, 14820.0, 57140.0, 38850.0, 22370.0, 62310.0, 670.0, 700.0], "Fyb": [333820.0, -61550.00000000001, 235639.99999999997, 21210.0, 118839.99999999999, -200139.99999999997, 98070.0, 65730.0, 37950.0, 52759.99999999999, -11560.0, 40690.0, -29500.0, 33200.0, -7720.000000000001, 15870.0, -11560.0, -6350.0, -14940.0, 8110.000000000001], "Fzb": [1008300.0, 985600.0, 946300.0, 895900.0, 837100.0, 770800.0, 699500.0, 624500.0, 549600.0, 474800.0, 401800.0, 331600.0, 267000.0, 207400.0, 154400.0, 107400.0, 69300.0, 37900.0, 17900.0, 3800.0], "Mxb": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], "Myb": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], "Mzb": [391100.0, 248970.0, 199030.0, 177620.0, 149070.0, 114450.00000000001, 74350.0, 27250.000000000004, -12750.0, -48040.0, -64720.0, -72200.0, -73190.0, -59870.0, -43390.0, -24650.0, -15320.0, -14680.000000000002, -13890.0, -8950.0], "Alpha": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], "prebend": [0.0112, 0.063, 0.1495, 0.2447, 0.3227, 0.3582, 0.3333, 0.2726, 0.188, 0.0609, -0.1235, -0.3422, -0.6135, -0.9416, -1.2977, -1.705, -2.1473, -2.6275, -3.1457, -3.708], "presweep": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]}
